Preprod

See merge request learnovia/learnovia-backend!727

+++ app/Helpers/CoursesHelper.php
@@ -4,20 +4,29 @@ namespace App\Helpers;
 use App\CourseItem;
 use App\UserCourseItem;
 use App\Material;
+use Modules\QuestionBank\Entities\Quiz;
 
 class CoursesHelper{
 
     public static function giveUsersAccessToViewCourseItem($itemId, $type , array $usersIds){
-        $item = CourseItem::create([
-            'item_id' => $itemId,
-            'type' => $type
-        ]);
 
-        foreach($usersIds as $userId){
-            UserCourseItem::create([
-                'user_id' => $userId,
-                'course_item_id' => $item->id,
+        $check=0;
+        if($type == 'quiz')
+            $check=Quiz::find($itemId)->draft;
+
+        if(!$check)
+        {
+            $item = CourseItem::firstOrCreate([
+                'item_id' => $itemId,
+                'type' => $type
             ]);
+    
+            foreach($usersIds as $userId){
+                UserCourseItem::firstOrCreate([
+                    'user_id' => $userId,
+                    'course_item_id' => $item->id,
+                ]);
+            }
         }
     }
 
@@ -25,9 +34,9 @@ class CoursesHelper{
         $courseItem = CourseItem::where('item_id', $itemId)->where('type', $type)->first();
         if(isset($usersIds)){
             if(isset($courseItem)){
-                $courseItem->courseItemUsers()->delete();
+                $courseItem->courseItemUsers()->whereNotIn('user_id',$usersIds)->delete();
                 foreach($usersIds as $userId){
-                    UserCourseItem::create([
+                    UserCourseItem::firstOrCreate([
                         'user_id' => $userId,
                         'course_item_id' => $courseItem->id,
                     ]);
